test(netcdf): split the 1799-line test_netcdf_unit.py by concern (Theme E.1)
Break the NetCDF unit-test module into five concern files plus a shared
helpers module, preserving all 81 tests (collection parity verified):

- test_netcdf_unit_properties.py (4 classes) - geotransform, caches,
  exception propagation, spatial delegates
- test_netcdf_unit_read.py (11 classes) - read_variable, read_md_array,
  needs_y_flip, and the get_variable edge/error/y-flip cases
- test_netcdf_unit_write.py (9 classes) - set/add/remove variable,
  metadata and no-data setters
- test_netcdf_unit_time.py (5 classes) - timestamp, get_time_variable
  (+units), cube dimension names, MSWEP real-world file
- test_netcdf_unit_create.py (5 classes) - create_from_array,
  validation, to_file, copy, read-file write mode

The four module-level _make_* builders move to a new shared
tests/netcdf/_netcdf_unit_helpers.py (the _make_3d_nc builder is used by
four of the five files); MSWEP_PATH is inlined into the time file. Each
file carries only the imports its classes use (isort-sectioned), with the
docstring-matched dead imports the AST splitter over-included removed.
