--- tests/netcdf/test_netcdf_unit.py
@@ -1,1799 +1,0 @@
-"""Unit tests targeting branch coverage in pyramids.netcdf.netcdf.NetCDF.
-
-Per-test docstrings describe the branch under test; pre-refactor
-line-number anchors were removed in favour of behavioural
-descriptions, which survive future code motion.
-
-Style: Google-style docstrings, <=120 char lines, no inline imports,
-descriptive assertion error messages.
-"""
-
-from __future__ import annotations
-
-from pathlib import Path
-from unittest.mock import MagicMock, PropertyMock, patch
-
-import numpy as np
-import pytest
-from osgeo import gdal
-
-from pyramids.dataset import Dataset
-from pyramids.netcdf.engines.variables import _create_netcdf_from_array
-from pyramids.netcdf.models import NetCDFMetadata
-from pyramids.netcdf.netcdf import NetCDF, Container
-from tests.netcdf.conftest import make_2d_nc, make_3d_nc
-
-pytestmark = pytest.mark.core
-
-
-def _make_3d_nc(
-    rows=10,
-    cols=12,
-    bands=3,
-    epsg=4326,
-    variable_name="temperature",
-    no_data_value=-9999.0,
-):
-    """Create a 3D in-memory NetCDF for testing.
-
-    Delegates to the shared ``make_3d_nc`` helper in conftest.
-    """
-    return make_3d_nc(
-        rows=rows,
-        cols=cols,
-        bands=bands,
-        epsg=epsg,
-        variable_name=variable_name,
-        no_data_value=no_data_value,
-        arr_type="random",
-        seed=42,
-    )
-
-
-def _make_dataset_2d(rows=10, cols=12, no_data=-9999.0):
-    """Create a 2D in-memory Dataset for testing.
-
-    Returns:
-        Dataset: A plain raster Dataset.
-    """
-    arr = np.random.RandomState(77).rand(rows, cols).astype(np.float64)
-    geo = (0.0, 1.0, 0, float(rows), 0, -1.0)
-    return Dataset.create_from_array(
-        arr,
-        geo=geo,
-        epsg=4326,
-        no_data_value=no_data,
-    )
-
-
-def _make_dataset_3d(bands=3, rows=10, cols=12, no_data=-9999.0):
-    """Create a 3D in-memory Dataset for testing.
-
-    Returns:
-        Dataset: A plain raster Dataset with multiple bands.
-    """
-    arr = np.random.RandomState(77).rand(bands, rows, cols).astype(np.float64)
-    geo = (0.0, 1.0, 0, float(rows), 0, -1.0)
-    return Dataset.create_from_array(
-        arr,
-        geo=geo,
-        epsg=4326,
-        no_data_value=no_data,
-    )
-
-
-class TestGeotransformFallback:
-    """Tests for geotransform property when lon/lat are unavailable."""
-
-    def test_geotransform_falls_back_when_lon_lat_none(self):
-        """Verify geotransform returns _geotransform when lon/lat are None.
-
-        Covers the branch returning self._geotransform when
-        lon and lat are not available from _read_variable.
-        """
-        nc = _make_3d_nc()
-        var = nc.get_variable("temperature")
-        # Patch lon and lat to return None to force fallback
-        with (
-            patch.object(
-                type(var), "lon", new_callable=PropertyMock, return_value=None
-            ),
-            patch.object(
-                type(var), "lat", new_callable=PropertyMock, return_value=None
-            ),
-        ):
-            gt = var.geotransform
-            assert gt is not None, "geotransform should not be None"
-            assert (
-                gt == var._geotransform
-            ), f"Expected _geotransform fallback {var._geotransform}, got {gt}"
-
-
-class TestInvalidateCaches:
-    """`_invalidate_caches` clears every per-instance cache (review L7)."""
-
-    def test_invalidate_clears_geostationary_gt_cache(self):
-        """`_invalidate_caches` empties the geostationary geotransform cache.
-
-        Test scenario:
-            The per-variable `_geostationary_gt_cache` is derived from the backing
-            geometry, so it must not survive a raster swap / in-place update.
-            Seed it, invalidate, and assert it (and the metadata cache) are cleared.
-        """
-        nc = make_2d_nc()
-        nc._geostationary_gt_cache["elevation"] = (0.0, 1.0, 0.0, 0.0, 0.0, -1.0)
-        nc._cached_meta_data = object()
-
-        nc._invalidate_caches()
-
-        assert nc._geostationary_gt_cache == {}, (
-            f"geostationary GT cache must be cleared, got {nc._geostationary_gt_cache}"
-        )
-        assert nc._cached_meta_data is None, "metadata cache must be cleared"
-
-
-class TestNarrowedExceptionPropagation:
-    """CON-9 narrowed `except RuntimeError` blocks let unexpected exceptions propagate (gap G9)."""
-
-    def _nc_with_fake_root_group(self, group_names_side_effect):
-        """A NetCDF whose root group's `GetGroupNames` raises the given exception.
-
-        Args:
-            group_names_side_effect: Exception instance to raise from `GetGroupNames`.
-
-        Returns:
-            NetCDF: An in-memory dataset with its `_raster` swapped for a mock whose
-            root group raises on `GetGroupNames`.
-        """
-        nc = make_2d_nc()
-        fake_rg = MagicMock()
-        fake_rg.GetGroupNames.side_effect = group_names_side_effect
-        fake_raster = MagicMock()
-        fake_raster.GetRootGroup.return_value = fake_rg
-        nc._raster = fake_raster
-        return nc
-
-    def test_group_names_propagates_unexpected_exception(self):
-        """An exception type outside the narrowed `except` propagates out of `group_names` (G9).
-
-        Test scenario:
-            CON-9 narrowed `group_names` to `except RuntimeError`. A `KeyError` from
-            `GetGroupNames` is *unexpected* and must now propagate (the whole point of the
-            narrowing) rather than being silently swallowed as it was under the old broad
-            `except Exception`.
-        """
-        nc = self._nc_with_fake_root_group(KeyError("boom"))
-        with pytest.raises(KeyError, match="boom"):
-            _ = nc.group_names
-
-    def test_group_names_degrades_on_runtime_error(self):
-        """An expected `RuntimeError` still degrades to an empty list (G9).
-
-        Test scenario:
-            The narrowed `except RuntimeError` keeps the graceful-degrade contract: a GDAL
-            `RuntimeError` from `GetGroupNames` yields `[]`, not a propagated error.
-        """
-        nc = self._nc_with_fake_root_group(RuntimeError("gdal driver error"))
-        assert nc.group_names == [], "RuntimeError should degrade to an empty group list"
-
-
-class TestNoDataValueSetter:
-    """Tests for NetCDF.no_data_value setter."""
-
-    def test_setter_with_single_value(self):
-        """Verify no_data_value setter handles a single scalar value.
-
-        Covers the else branch that calls
-        _change_no_data_value_attr(0, value) for a scalar.
-        """
-        nc = make_2d_nc()
-        var = nc.get_variable("elevation")
-        var.no_data_value = -1.0
-        assert (
-            var.no_data_value[0] == -1.0
-        ), f"Expected -1.0, got {var.no_data_value[0]}"
-
-    def test_setter_with_list_value(self):
-        """Verify no_data_value setter handles a list of values.
-
-        Covers the if-isinstance(value, list) branch
-        that iterates and sets per-band no-data values.
-        """
-        nc = _make_3d_nc()
-        var = nc.get_variable("temperature")
-        new_values = [-1.0, -2.0, -3.0]
-        var.no_data_value = new_values
-        for i, expected in enumerate(new_values):
-            assert (
-                var.no_data_value[i] == expected
-            ), f"Band {i}: expected {expected}, got {var.no_data_value[i]}"
-
-
-class TestTimeStamp:
-    """Tests for NetCDF.time_stamp property."""
-
-    def test_time_stamp_returns_none_without_time_units(self):
-        """Verify time_stamp returns None when there is no time units attribute.
-
-        Covers delegates to get_time_variable() which returns
-        None when time dimension lacks a 'units' attribute.
-        """
-        nc = _make_3d_nc()
-        result = nc.time_stamp
-        assert (
-            result is None
-        ), f"Expected None (no time units in created NC), got {result}"
-
-
-class TestGetTimeVariable:
-    """Tests for NetCDF.get_time_variable method."""
-
-    def test_get_time_variable_with_units(self):
-        """Verify get_time_variable parses time when units attribute exists.
-
-        Covers the full path through get_time_variable
-        where time_dim has units and time values can be converted.
-        """
-        nc = NetCDF.read_file(
-            "tests/data/netcdf/noah-precipitation-1979.nc",
-            open_as_multi_dimensional=True,
-        )
-        result = nc.get_time_variable()
-        if result is not None:
-            assert isinstance(result, list), f"Expected list, got {type(result)}"
-            assert len(result) > 0, "Expected non-empty time list"
-        # If no time units in this file, it will be None and that's OK
-
-    def test_get_time_variable_no_time_dim(self):
-        """Verify get_time_variable returns None when no time dimension exists.
-
-        Covers (time_stamp = None) return.
-        """
-        nc = make_2d_nc()
-        result = nc.get_time_variable()
-        assert (
-            result is None
-        ), f"Expected None for NC without time dimension, got {result}"
-
-    def test_get_time_variable_custom_format(self):
-        """Verify get_time_variable respects custom time_format.
-
-        Covers the conversion path with a custom format.
-        """
-        nc = NetCDF.read_file(
-            "tests/data/netcdf/noah-precipitation-1979.nc",
-            open_as_multi_dimensional=True,
-        )
-        result = nc.get_time_variable(time_format="%Y/%m/%d")
-        if result is not None:
-            assert "/" in result[0], f"Expected '/' in date format, got {result[0]}"
-
-    def test_get_time_variable_cds_beta_era5(self):
-        """E2E: CDS-Beta ERA5 NetCDFs hide ``valid_time#units`` from the
-        multidim driver but expose it on the classic driver. The
-        ``MetadataBuilder._topup_dim_attrs_from_classic`` fallback must
-        recover ``units`` (and ``calendar``) so ``get_time_variable``
-        returns parsed dates instead of ``None``. Reproduces issue #309.
-        """
-        nc = NetCDF.read_file(
-            "tests/data/netcdf/era5_cds_beta_t2m_jan2022.nc",
-            open_as_multi_dimensional=True,
-        )
-
-        valid_time = nc.meta_data.dimensions["valid_time"]
-        assert (
-            valid_time.attrs.get("units") == "seconds since 1970-01-01"
-        ), f"expected classic units to be merged, got attrs={valid_time.attrs}"
-        assert (
-            valid_time.attrs.get("calendar") == "proleptic_gregorian"
-        ), f"expected calendar from multidim to survive, got attrs={valid_time.attrs}"
-
-        result = nc.get_time_variable("valid_time")
-        assert result is not None, (
-            "get_time_variable returned None on a CDS-Beta NetCDF — the "
-            "classic-metadata top-up did not surface 'units' on valid_time"
-        )
-        assert isinstance(result, list), f"expected list, got {type(result)}"
-        assert (
-            len(result) == valid_time.size
-        ), f"expected {valid_time.size} timestamps, got {len(result)}"
-        assert result[0].startswith(
-            "2022-01-"
-        ), f"expected 2022-01-* from the CDS-Beta sample, got {result[0]}"
-
-
-class TestSpatialOperationDelegates:
-    """Tests for crop() and to_crs() delegation to parent class."""
-
-    def test_crop_delegates_to_super(self):
-        """Verify crop() passes through to Dataset.crop for subsets.
-
-        Covers super().crop() call after _check_not_container.
-        """
-        nc = _make_3d_nc(rows=20, cols=24, bands=2)
-        var = nc.get_variable("temperature")
-        import geopandas as gpd
-        from shapely.geometry import box
-
-        mask = gpd.GeoDataFrame(
-            geometry=[box(1.0, 1.0, 5.0, 5.0)],
-            crs="EPSG:4326",
-        )
-        result = var.crop(mask, touch=True)
-        assert result is not None, "crop should return a new Dataset"
-        assert (
-            result.rows <= var.rows
-        ), f"Cropped rows {result.rows} should be <= original {var.rows}"
-
-    def test_to_crs_delegates_to_super(self):
-        """Verify to_crs() passes through to Dataset.to_crs for subsets.
-
-        Covers super().to_crs() call after _check_not_container.
-        """
-        nc = _make_3d_nc(rows=10, cols=12, bands=1, epsg=4326)
-        var = nc.get_variable("temperature")
-        result = var.to_crs(to_epsg=32637)
-        assert result is not None, "to_crs should return a reprojected Dataset"
-
-
-class TestReadFileWriteMode:
-    """Tests for NetCDF.read_file with read_only=False."""
-
-    def test_read_file_write_mode(self, tmp_path):
-        """Verify read_file with read_only=False opens in write mode.
-
-        Covers the else branch setting read_only = 'write'.
-        """
-        nc = make_2d_nc()
-        out = str(tmp_path / "writable.nc")
-        nc.to_file(out)
-        writable_nc = NetCDF.read_file(
-            out,
-            read_only=False,
-            open_as_multi_dimensional=True,
-        )
-        assert writable_nc is not None, "Should open file for writing"
-        assert (
-            writable_nc._access == "write"
-        ), f"Expected 'write' access, got {writable_nc._access}"
-
-
-class TestMetaDataSetter:
-    """Tests for NetCDF.meta_data setter."""
-
-    def test_setter_with_dict(self):
-        """Verify meta_data setter accepts a plain dict and sets items.
-
-        Covers the isinstance(value, dict) branch
-        calling SetMetadataItem for each key.
-        """
-        nc = make_2d_nc()
-        nc.meta_data = {"source": "test", "version": "1.0"}
-        gdal_meta = nc._raster.GetMetadata()
-        assert (
-            gdal_meta.get("source") == "test"
-        ), f"Expected 'test', got {gdal_meta.get('source')}"
-
-    def test_setter_with_netcdf_metadata(self):
-        """Verify meta_data setter accepts a NetCDFMetadata object.
-
-        Covers the else branch that directly sets
-        _cached_meta_data.
-        """
-        from pyramids.netcdf.models import DimensionInfo
-
-        nc = make_2d_nc()
-        custom_meta = NetCDFMetadata(
-            driver="netCDF",
-            root_group="/",
-            groups={},
-            variables={},
-            dimensions={
-                "/x": DimensionInfo(name="x", full_name="/x", size=12),
-            },
-            global_attributes={},
-            structural=None,
-            created_with={"gdal": "3.12"},
-        )
-        nc.meta_data = custom_meta
-        assert (
-            nc._cached_meta_data is custom_meta
-        ), "Expected _cached_meta_data to be the assigned object"
-
-
-class TestReadVariable:
-    """Tests for NetCDF._read_variable private method."""
-
-    def test_read_variable_dimension_indexing_variable(self):
-        """Verify _read_variable reads coordinate arrays via dimension indexing.
-
-        Covers the fallback to dim.GetIndexingVariable()
-        when OpenMDArray returns None.
-        """
-        nc = make_2d_nc()
-        result = nc._read_variable("x")
-        assert result is not None, "Should read 'x' dimension values"
-        assert isinstance(
-            result, np.ndarray
-        ), f"Expected np.ndarray, got {type(result)}"
-
-    def test_read_variable_classic_mode(self):
-        """Verify _read_variable works in classic mode via subdataset string.
-
-        Covers the classic-mode branch that opens via
-        gdal.Open(f'NETCDF:{path}:{var}').
-        """
-        nc = NetCDF.read_file(
-            "tests/data/netcdf/noah-precipitation-1979.nc",
-            open_as_multi_dimensional=False,
-        )
-        var_names = nc.variable_names
-        if var_names:
-            result = nc._read_variable(var_names[0])
-            assert (
-                result is not None
-            ), f"Should read variable '{var_names[0]}' in classic mode"
-
-    def test_read_variable_nonexistent_returns_none(self):
-        """Verify _read_variable returns None for nonexistent variables."""
-        nc = make_2d_nc()
-        result = nc._read_variable("nonexistent_variable_xyz")
-        assert result is None, f"Expected None for nonexistent variable, got {result}"
-
-    def test_read_variable_classic_mode_nonexistent(self):
-        """Verify _read_variable returns None in classic mode for bad var name.
-
-        Covers the except (RuntimeError, AttributeError) in classic mode.
-        """
-        nc = NetCDF.read_file(
-            "tests/data/netcdf/noah-precipitation-1979.nc",
-            open_as_multi_dimensional=False,
-        )
-        result = nc._read_variable("totally_fake_var")
-        assert result is None, "Expected None for nonexistent var in classic mode"
-
-
-class TestReadMdArray1D:
-    """Tests for _read_md_array with 1D variables."""
-
-    def test_read_md_array_1d_string_type(self):
-        """Verify _read_md_array handles 1D string-typed arrays.
-
-        Covers the len(dims)==1 branch with a
-        GEDTC_STRING dtype, returning (md_arr, md_arr, rg).
-        """
-        # Create an MDIM dataset with a 1D string variable
-        src_ds = gdal.GetDriverByName("MEM").CreateMultiDimensional("test")
-        rg = src_ds.GetRootGroup()
-        dim = rg.CreateDimension("labels", None, None, 3)
-        str_dtype = gdal.ExtendedDataType.CreateString()
-        str_arr = rg.CreateMDArray("label_data", [dim], str_dtype)
-        nc = Container(src_ds)
-        result_src, result_md, result_rg, _ix, _iy = nc._read_md_array("label_data")
-        # For string type, src should be the md_arr itself (not a Dataset)
-        assert (
-            result_src is result_md
-        ), "For string 1D arrays, src and md_arr should be the same object"
-        assert result_rg is not None, "root group ref should not be None"
-
-
-class TestNeedsYFlip:
-    """Tests for the _needs_y_flip instance method."""
-
-    def test_returns_false_for_1d_array(self):
-        """Verify _needs_y_flip returns False for 1-D arrays.
-
-        1-D arrays have no Y axis to flip.
-        """
-        nc = make_2d_nc()
-        rg = nc._raster.GetRootGroup()
-        md_arr = rg.OpenMDArray("x")
-        result = nc._needs_y_flip(rg, md_arr)
-        assert result is False, f"Expected False for 1-D array, got {result}"
-
-    def test_returns_bool_for_2d_array(self):
-        """Verify _needs_y_flip returns a bool for 2D arrays.
-
-        The result depends on whether the Y dimension goes
-        south-to-north (positive Y pixel size) or not. For
-        in-memory datasets created by pyramids the geotransform
-        is already GDAL-convention (negative Y), so the result
-        may be False.
-        """
-        nc = make_2d_nc()
-        rg = nc._raster.GetRootGroup()
-        md_arr = rg.OpenMDArray("elevation")
-        result = nc._needs_y_flip(rg, md_arr)
-        assert isinstance(result, bool), f"Expected bool, got {type(result)}"
-
-
-class TestGetVariableEdgeCases:
-    """Tests for get_variable edge cases."""
-
-    def test_get_variable_invalid_name_raises(self):
-        """Verify get_variable raises ValueError for invalid variable name.
-
-        Covers branch where src is None after gdal.Open.
-        """
-        nc = _make_3d_nc()
-        with pytest.raises(ValueError, match="not a valid variable name"):
-            nc.get_variable("nonexistent_variable")
-
-    def test_get_variable_classic_mode(self):
-        """Verify get_variable works in classic mode (no root group).
-
-        Covers the else branch using
-        NETCDF:file:variable_name.
-        """
-        nc = NetCDF.read_file(
-            "tests/data/netcdf/noah-precipitation-1979.nc",
-            open_as_multi_dimensional=False,
-        )
-        var_names = nc.variable_names
-        if var_names:
-            var = nc.get_variable(var_names[0])
-            assert var.is_subset is True, "Variable should be a subset"
-            assert (
-                var._is_md_array is False
-            ), "Classic-mode variable should not be md_array"
-
-    def test_get_variable_sets_md_array_dims(self):
-        """Verify get_variable populates _md_array_dims.
-
-        Covers the code where dims are stored.
-        """
-        nc = _make_3d_nc()
-        var = nc.get_variable("temperature")
-        assert isinstance(
-            var._md_array_dims, list
-        ), f"Expected list, got {type(var._md_array_dims)}"
-        assert (
-            len(var._md_array_dims) == 3
-        ), f"Expected 3 dims, got {len(var._md_array_dims)}"
-
-    def test_get_variable_sets_band_dim_info(self):
-        """Verify get_variable populates _band_dim_name and _band_dim_values.
-
-        Covers where band dimension info is extracted.
-        """
-        nc = _make_3d_nc()
-        var = nc.get_variable("temperature")
-        assert (
-            var._band_dim_name is not None
-        ), "Expected a band dim name for 3D variable"
-        assert (
-            var._band_dim_values is not None
-        ), "Expected band dim values for 3D variable"
-        assert (
-            len(var._band_dim_values) == 3
-        ), f"Expected 3 band values, got {len(var._band_dim_values)}"
-
-    def test_get_variable_2d_has_no_band_dim(self):
-        """Verify get_variable sets _band_dim_name=None for 2D variables.
-
-        Covers the else branch where ndim <= 2.
-        """
-        nc = make_2d_nc()
-        var = nc.get_variable("elevation")
-        assert (
-            var._band_dim_name is None
-        ), f"Expected None band_dim_name for 2D var, got {var._band_dim_name}"
-        assert (
-            var._band_dim_values is None
-        ), f"Expected None band_dim_values for 2D var, got {var._band_dim_values}"
-
-
-class TestGetVariableBandDimErrors:
-    """Tests for get_variable band dimension error paths."""
-
-    def test_get_variable_band_dim_runtime_error_fallback(self):
-        """Verify get_variable handles RuntimeError when reading band dim values.
-
-        Covers when ReadAsArray on the indexing variable
-        raises RuntimeError, falls back to range indices.
-        """
-        nc = _make_3d_nc()
-        var = nc.get_variable("temperature")
-        # The band dim values should be populated even in normal case
-        assert var._band_dim_values is not None, "Band dim values should not be None"
-
-
-class TestToFile:
-    """Tests for NetCDF.to_file edge cases."""
-
-    def test_to_file_tif_extension_for_subset(self, tmp_path):
-        """Verify to_file works for non-.nc extensions on variable subsets.
-
-        Covers super().to_file() path for subsets.
-        """
-        nc = make_2d_nc()
-        var = nc.get_variable("elevation")
-        out = tmp_path / "output.tif"
-        var.to_file(out)
-        assert out.exists(), f"File should exist at {out}"
-        assert out.stat().st_size > 0, "File should not be empty"
-
-    def test_to_file_non_nc_on_container_raises(self, tmp_path):
-        """Verify to_file raises ValueError for non-.nc on root containers.
-
-        Covers the ValueError for multidimensional
-        container + non-nc extension.
-        """
-        nc = make_2d_nc()
-        out = tmp_path / "output.tif"
-        with pytest.raises(ValueError, match="Cannot save a multidimensional"):
-            nc.to_file(out)
-
-    def test_to_file_nc_creates_copy_failure(self, tmp_path):
-        """Verify to_file raises RuntimeError when CreateCopy fails.
-
-        Covers the RuntimeError branch.
-        """
-        nc = make_2d_nc()
-        with patch.object(gdal.Driver, "CreateCopy", return_value=None):
-            out = str(tmp_path / "bad_output.nc")
-            with pytest.raises(RuntimeError, match="Failed to save NetCDF"):
-                nc.to_file(out)
-
-
-class TestCopy:
-    """Tests for NetCDF.copy edge cases."""
-
-    def test_copy_failure_raises_runtime_error(self):
-        """Verify copy raises RuntimeError when CreateCopy fails.
-
-        Covers the RuntimeError branch.
-        """
-        nc = make_2d_nc()
-        with patch.object(gdal.Driver, "CreateCopy", return_value=None):
-            with pytest.raises(RuntimeError, match="Failed to copy"):
-                nc.copy()
-
-    def test_copy_to_file_path(self, tmp_path):
-        """Verify copy with a file path uses netCDF driver.
-
-        Covers the else branch setting driver='netCDF'.
-        """
-        nc = make_2d_nc()
-        out = tmp_path / "copy_output.nc"
-        copied = nc.copy(path=out)
-        assert copied is not None, "Copy should return a valid NetCDF"
-        assert out.exists(), f"File should exist at {out}"
-
-
-class TestCreateFromArrayAlternatives:
-    """Tests for create_from_array alternative parameter paths."""
-
-    def test_create_from_array_with_top_left_and_cell_size(self):
-        """Verify create_from_array builds geo from top_left_corner and cell_size.
-
-        Covers the branch building geo from
-        top_left_corner and cell_size.
-        """
-        arr = np.random.rand(5, 10).astype(np.float64)
-        nc = NetCDF.create_from_array(
-            arr=arr,
-            top_left_corner=(10.0, 50.0),
-            cell_size=0.5,
-            epsg=4326,
-            no_data_value=-9999.0,
-            path=None,
-        )
-        assert nc is not None, "NetCDF should be created"
-        var = nc.get_variable("data")
-        assert var.cell_size == pytest.approx(0.5), f"Expected cell_size 0.5, got {var.cell_size}"
-
-    def test_create_from_array_no_geo_raises(self):
-        """Verify create_from_array raises ValueError without geo information.
-
-        Covers the ValueError when geo is None and
-        top_left_corner/cell_size are not both provided.
-        """
-        arr = np.random.rand(5, 10).astype(np.float64)
-        with pytest.raises(ValueError, match="Either 'geo'"):
-            NetCDF.create_from_array(
-                arr=arr,
-                epsg=4326,
-                no_data_value=-9999.0,
-            )
-
-    def test_create_from_array_default_variable_name(self):
-        """Verify create_from_array defaults variable_name to 'data'.
-
-        Covers variable_name = 'data' default.
-        """
-        arr = np.random.rand(5, 10).astype(np.float64)
-        geo = (0.0, 1.0, 0, 5.0, 0, -1.0)
-        nc = NetCDF.create_from_array(
-            arr=arr,
-            geo=geo,
-            epsg=4326,
-            no_data_value=-9999.0,
-            path=None,
-        )
-        assert (
-            "data" in nc.variable_names
-        ), f"Expected 'data' in variable_names, got {nc.variable_names}"
-
-    def test_create_from_array_default_extra_dim_values(self):
-        """Verify create_from_array defaults extra_dim_values to 0..N-1.
-
-        Covers the default extra_dim_values generation for 3D arrays.
-        """
-        arr = np.random.rand(4, 5, 10).astype(np.float64)
-        geo = (0.0, 1.0, 0, 5.0, 0, -1.0)
-        nc = NetCDF.create_from_array(
-            arr=arr,
-            geo=geo,
-            epsg=4326,
-            no_data_value=-9999.0,
-            variable_name="test_var",
-            path=None,
-        )
-        var = nc.get_variable("test_var")
-        assert var.band_count == 4, f"Expected 4 bands, got {var.band_count}"
-
-
-class TestCreateNetcdfFromArrayValidation:
-    """Tests for _create_netcdf_from_array input validation."""
-
-    def test_variable_name_none_raises(self):
-        """Verify _create_netcdf_from_array raises ValueError for None variable_name.
-
-        Covers the ValueError for variable_name is None.
-        """
-        arr = np.random.rand(5, 10).astype(np.float64)
-        with pytest.raises(ValueError, match="Variable_name cannot be None"):
-            _create_netcdf_from_array(
-                arr,
-                None,
-                10,
-                5,
-                geo=(0.0, 1.0, 0, 5.0, 0, -1.0),
-            )
-
-    def test_geo_none_raises(self):
-        """Verify _create_netcdf_from_array raises ValueError for None geo.
-
-        Covers the ValueError for geo is None.
-        """
-        arr = np.random.rand(5, 10).astype(np.float64)
-        with pytest.raises(ValueError, match="geo cannot be None"):
-            _create_netcdf_from_array(
-                arr,
-                "var",
-                10,
-                5,
-                geo=None,
-            )
-
-
-class TestAddMdArrayToGroupFallback:
-    """Tests for _add_md_array_to_group NoData handling."""
-
-    def test_no_nodata_when_source_has_none(self):
-        """When source has no nodata, the copy should also have no nodata.
-
-        Test scenario:
-            Source variable with GetNoDataValue() returning None should
-            not produce a phantom -9999 sentinel on the copy.
-        """
-        nc = make_2d_nc()
-        src_rg = nc._raster.GetRootGroup()
-        src_arr = src_rg.OpenMDArray("elevation")
-
-        dst = gdal.GetDriverByName("MEM").CreateMultiDimensional("dst")
-        dst_rg = dst.GetRootGroup()
-        dtype = gdal.ExtendedDataType.Create(gdal.GDT_Float64)
-        for d in src_arr.GetDimensions():
-            iv = d.GetIndexingVariable()
-            NetCDF.create_main_dimension(dst_rg, d.GetName(), dtype, iv.ReadAsArray())
-
-        # Patch GetNoDataValue to return None (no nodata on source)
-        with patch.object(type(src_arr), "GetNoDataValue", return_value=None):
-            NetCDF._add_md_array_to_group(dst_rg, "copied_var", src_arr)
-
-        copied = dst_rg.OpenMDArray("copied_var")
-        assert copied is not None, "Copied variable should exist"
-        ndv = copied.GetNoDataValue()
-        assert ndv is None, f"Expected no nodata (None), got {ndv}"
-
-
-class TestSetVariableAttributes:
-    """Tests for set_variable attribute writing paths."""
-
-    def test_set_variable_with_float_attr(self):
-        """Verify set_variable writes float attributes.
-
-        Covers the float attribute branch.
-        """
-        nc = make_2d_nc()
-        ds = _make_dataset_2d()
-        nc.set_variable(
-            "pressure",
-            ds,
-            attrs={"scale_factor": 1.5},
-        )
-        rg = nc._raster.GetRootGroup()
-        md_arr = rg.OpenMDArray("pressure")
-        attr_names = [a.GetName() for a in md_arr.GetAttributes()]
-        assert (
-            "scale_factor" in attr_names
-        ), f"Expected 'scale_factor' attribute, got {attr_names}"
-
-    def test_set_variable_with_int_attr(self):
-        """Verify set_variable writes integer attributes.
-
-        Covers the int attribute branch.
-        """
-        nc = make_2d_nc()
-        ds = _make_dataset_2d()
-        nc.set_variable(
-            "pressure",
-            ds,
-            attrs={"flag": 42},
-        )
-        rg = nc._raster.GetRootGroup()
-        md_arr = rg.OpenMDArray("pressure")
-        attr_names = [a.GetName() for a in md_arr.GetAttributes()]
-        assert "flag" in attr_names, f"Expected 'flag' attribute, got {attr_names}"
-
-    def test_set_variable_with_non_string_non_numeric_attr(self):
-        """Verify set_variable converts unknown types to string.
-
-        Covers the else branch converting value to
-        str and using CreateString.
-        """
-        nc = make_2d_nc()
-        ds = _make_dataset_2d()
-        nc.set_variable(
-            "pressure",
-            ds,
-            attrs={"metadata": [1, 2, 3]},
-        )
-        rg = nc._raster.GetRootGroup()
-        md_arr = rg.OpenMDArray("pressure")
-        assert md_arr is not None, "pressure variable should exist"
-
-    def test_set_variable_with_string_attr(self):
-        """Verify set_variable writes string attributes.
-
-        Covers the string attribute branch.
-        """
-        nc = make_2d_nc()
-        ds = _make_dataset_2d()
-        nc.set_variable(
-            "wind",
-            ds,
-            attrs={"units": "m/s"},
-        )
-        rg = nc._raster.GetRootGroup()
-        md_arr = rg.OpenMDArray("wind")
-        attr_names = [a.GetName() for a in md_arr.GetAttributes()]
-        assert "units" in attr_names, f"Expected 'units' attribute, got {attr_names}"
-
-    def test_set_variable_no_data_exception_path(self):
-        """Verify set_variable handles exception when SetNoDataValueDouble fails.
-
-        Covers the except branch in no-data setting.
-        """
-        nc = make_2d_nc()
-        ds = _make_dataset_2d()
-        # The normal path should set no data without error
-        nc.set_variable("with_nodata", ds)
-        rg = nc._raster.GetRootGroup()
-        md_arr = rg.OpenMDArray("with_nodata")
-        assert md_arr is not None, "Variable should exist"
-
-    def test_set_variable_replaces_existing(self):
-        """Verify set_variable deletes and replaces an existing variable.
-
-        Covers rg.DeleteMDArray(variable_name).
-        """
-        nc = make_2d_nc()
-        ds1 = _make_dataset_2d()
-        ds2 = _make_dataset_2d(rows=10, cols=12)
-        nc.set_variable("replace_me", ds1)
-        assert (
-            "replace_me" in nc.variable_names
-        ), "Variable should exist before replacement"
-        nc.set_variable("replace_me", ds2)
-        assert (
-            "replace_me" in nc.variable_names
-        ), "Variable should still exist after replacement"
-
-    def test_set_variable_3d_with_no_band_dim(self):
-        """Verify set_variable auto-names band dim as 'bands'.
-
-        Covers default band_dim_name and values.
-        """
-        nc = make_2d_nc()
-        ds = _make_dataset_3d(bands=2, rows=10, cols=12)
-        nc.set_variable("multi_band", ds)
-        rg = nc._raster.GetRootGroup()
-        md_arr = rg.OpenMDArray("multi_band")
-        dims = md_arr.GetDimensions()
-        assert len(dims) == 3, f"Expected 3 dims for 3D var, got {len(dims)}"
-        dim_names = [d.GetName() for d in dims]
-        assert "bands" in dim_names or any(
-            "band" in n for n in dim_names
-        ), f"Expected a 'bands' dimension, got {dim_names}"
-
-    def test_set_variable_attr_exception_silenced(self):
-        """Verify set_variable silences exceptions when writing attributes.
-
-        Covers the except pass block.
-        """
-        nc = make_2d_nc()
-        ds = _make_dataset_2d()
-        # This should not raise even if CreateAttribute fails internally
-        nc.set_variable(
-            "safe_var",
-            ds,
-            attrs={"units": "K", "count": 5, "ratio": 3.14, "complex": [1, 2]},
-        )
-        rg = nc._raster.GetRootGroup()
-        md_arr = rg.OpenMDArray("safe_var")
-        assert md_arr is not None, "Variable should exist"
-
-    def test_set_variable_without_root_group_raises(self):
-        """Verify set_variable raises ValueError when no root group.
-
-        Covers .
-        """
-        nc = NetCDF.read_file(
-            "tests/data/netcdf/noah-precipitation-1979.nc",
-            open_as_multi_dimensional=False,
-        )
-        ds = _make_dataset_2d()
-        with pytest.raises(ValueError, match="set_variable requires"):
-            nc.set_variable("new_var", ds)
-
-
-class TestAddVariable:
-    """Tests for add_variable edge cases."""
-
-    def test_add_variable_with_specific_name(self):
-        """Verify add_variable copies a specific variable by name.
-
-        Covers names_to_copy = [variable_name].
-        """
-        nc = _make_3d_nc(variable_name="temp")
-        nc2 = _make_3d_nc(variable_name="precip")
-        nc.add_variable(nc2, variable_name="precip")
-        assert (
-            "precip" in nc.variable_names
-        ), f"Expected 'precip' in {nc.variable_names}"
-
-    def test_add_variable_non_netcdf_dataset(self):
-        """Verify add_variable with a plain Dataset gives empty names_to_copy.
-
-        Covers names_to_copy = [] for non-NetCDF dataset.
-        """
-        nc = _make_3d_nc(variable_name="temp")
-        ds = _make_dataset_2d()
-        # Assign a _raster with a root group via a mock
-        original_names = nc.variable_names[:]
-        # This should not raise and should not change variable names
-        # because names_to_copy will be []
-        mock_rg = MagicMock()
-        mock_rg.OpenMDArray = MagicMock(return_value=None)
-        ds._raster = MagicMock()
-        ds._raster.GetRootGroup.return_value = mock_rg
-        nc.add_variable(ds)
-        # Variable names should not change since names_to_copy is empty
-        assert (
-            nc.variable_names == original_names
-        ), f"Variable names should not change, got {nc.variable_names}"
-
-
-class TestRemoveVariable:
-    """Tests for remove_variable on non-memory datasets."""
-
-    def test_remove_variable_from_file_based_dataset(self, tmp_path):
-        """Verify remove_variable copies to memory for file-based datasets.
-
-        Covers the else branch using CreateCopy for
-        non-memory drivers.
-        """
-        nc = _make_3d_nc(variable_name="temp")
-        out = str(tmp_path / "to_remove.nc")
-        nc.to_file(out)
-        file_nc = NetCDF.read_file(
-            out,
-            read_only=False,
-            open_as_multi_dimensional=True,
-        )
-        assert (
-            "temp" in file_nc.variable_names
-        ), "Variable 'temp' should exist before removal"
-        file_nc.remove_variable("temp")
-        assert "temp" not in file_nc.variable_names, "Variable 'temp' should be removed"
-
-    def test_remove_variable_in_memory(self):
-        """Verify remove_variable works directly for in-memory datasets.
-
-        Covers the if driver_type == 'memory' branch.
-        """
-        nc = _make_3d_nc(variable_name="temp")
-        assert "temp" in nc.variable_names, "Variable should exist before removal"
-        nc.remove_variable("temp")
-        assert (
-            "temp" not in nc.variable_names
-        ), "Variable should be removed from in-memory dataset"
-
-
-MSWEP_PATH = Path("tests/data/netcdf/MSWEP_1979010100.nc")
-
-
-class TestMSWEPFile:
-    """Tests using the MSWEP test file for real-world coverage.
-
-    ``MSWEP_PATH`` is committed test data; if it is ever missing these tests
-    must fail loudly rather than skip silently, so there is no ``.exists()`` gate.
-    """
-
-    def test_read_mswep_mdim(self):
-        """Verify reading MSWEP file in multidimensional mode.
-
-        Uses tests/data/netcdf/MSWEP_1979010100.nc to hit real code paths.
-        """
-        nc = NetCDF.read_file(
-            str(MSWEP_PATH),
-            open_as_multi_dimensional=True,
-        )
-        assert (
-            "precipitation" in nc.variable_names
-        ), f"Expected 'precipitation' in {nc.variable_names}"
-        var = nc.get_variable("precipitation")
-        assert var.is_subset is True, "Variable should be a subset"
-        arr = var.read_array()
-        assert arr is not None, "Should read array from variable"
-        assert arr.ndim >= 2, f"Expected 2D+ array, got {arr.ndim}D"
-
-    def test_mswep_dimension_names(self):
-        """Verify MSWEP file has correct dimension names."""
-        nc = NetCDF.read_file(
-            str(MSWEP_PATH),
-            open_as_multi_dimensional=True,
-        )
-        dims = nc.dimension_names
-        assert dims is not None, "Dimension names should not be None"
-        assert "lon" in dims, f"Expected 'lon' in {dims}"
-        assert "lat" in dims, f"Expected 'lat' in {dims}"
-
-    def test_mswep_meta_data(self):
-        """Verify MSWEP metadata is accessible."""
-        nc = NetCDF.read_file(
-            str(MSWEP_PATH),
-            open_as_multi_dimensional=True,
-        )
-        md = nc.meta_data
-        assert isinstance(
-            md, NetCDFMetadata
-        ), f"Expected NetCDFMetadata, got {type(md)}"
-
-    def test_mswep_get_all_metadata(self):
-        """Verify get_all_metadata populates dimension overview."""
-        nc = NetCDF.read_file(
-            str(MSWEP_PATH),
-            open_as_multi_dimensional=True,
-        )
-        md = nc.get_all_metadata()
-        assert len(md.dimensions) > 0, "dimensions should be populated"
-
-    def test_mswep_lon_lat(self):
-        """Verify lon/lat are readable from MSWEP file."""
-        nc = NetCDF.read_file(
-            str(MSWEP_PATH),
-            open_as_multi_dimensional=True,
-        )
-        lon = nc.lon
-        lat = nc.lat
-        assert lon is not None, "lon should not be None"
-        assert lat is not None, "lat should not be None"
-        assert lon.ndim == 1, f"lon should be 1D, got {lon.ndim}D"
-        assert lat.ndim == 1, f"lat should be 1D, got {lat.ndim}D"
-
-
-def _make_nc_with_time_units(rows=4, cols=5, n_times=3):
-    """Create an MDIM NetCDF with a time dimension that has a 'units' attr.
-
-    This is needed to exercise the get_time_variable conversion path.
-
-    Returns:
-        NetCDF: An in-memory NetCDF with a time dimension carrying
-            a ``units`` attribute of ``"days since 1979-01-01"``.
-    """
-    src = gdal.GetDriverByName("MEM").CreateMultiDimensional("time_test")
-    rg = src.GetRootGroup()
-    dtype = gdal.ExtendedDataType.Create(gdal.GDT_Float64)
-
-    # Create x dimension
-    dim_x = rg.CreateDimension("x", "HORIZONTAL_X", None, cols)
-    x_vals = rg.CreateMDArray("x", [dim_x], dtype)
-    x_vals.Write(np.arange(cols, dtype=np.float64) + 0.5)
-    dim_x.SetIndexingVariable(x_vals)
-
-    # Create y dimension
-    dim_y = rg.CreateDimension("y", "HORIZONTAL_Y", None, rows)
-    y_vals = rg.CreateMDArray("y", [dim_y], dtype)
-    y_vals.Write(np.arange(rows, dtype=np.float64)[::-1] + 0.5)
-    dim_y.SetIndexingVariable(y_vals)
-
-    # Create time dimension with units attribute
-    dim_t = rg.CreateDimension("time", "TEMPORAL", None, n_times)
-    t_vals = rg.CreateMDArray("time", [dim_t], dtype)
-    t_vals.Write(np.arange(n_times, dtype=np.float64))
-    dim_t.SetIndexingVariable(t_vals)
-
-    # Add 'units' attribute to the time variable
-    str_dtype = gdal.ExtendedDataType.CreateString()
-    units_attr = t_vals.CreateAttribute("units", [], str_dtype)
-    units_attr.Write("days since 1979-01-01")
-
-    # Create a data variable
-    data_arr = rg.CreateMDArray("temperature", [dim_t, dim_y, dim_x], dtype)
-    data_arr.Write(np.random.RandomState(55).rand(n_times, rows, cols))
-    data_arr.SetNoDataValueDouble(-9999.0)
-
-    return Container(src)
-
-
-class TestGetTimeVariableWithUnits:
-    """Tests for get_time_variable with actual time units attribute."""
-
-    def test_get_time_variable_with_days_since(self):
-        """Verify get_time_variable converts time values when units exist.
-
-        Covers the full path where time_dim has units,
-        time values are read, and the conversion function is applied.
-        """
-        nc = _make_nc_with_time_units(n_times=3)
-        result = nc.get_time_variable()
-        assert (
-            result is not None
-        ), "get_time_variable should return dates when units exist"
-        assert isinstance(result, list), f"Expected list, got {type(result)}"
-        assert len(result) == 3, f"Expected 3 time stamps, got {len(result)}"
-        assert (
-            "1979-01-01" in result[0]
-        ), f"Expected '1979-01-01' in first timestamp, got {result[0]}"
-        assert (
-            "1979-01-02" in result[1]
-        ), f"Expected '1979-01-02' in second timestamp, got {result[1]}"
-
-    def test_get_time_variable_custom_format(self):
-        """Verify get_time_variable uses a custom format string.
-
-        Covers the create_time_conversion_func call with
-        custom time_format.
-        """
-        nc = _make_nc_with_time_units(n_times=2)
-        result = nc.get_time_variable(time_format="%Y/%m/%d")
-        assert result is not None, "Should return formatted timestamps"
-        assert (
-            "/" in result[0]
-        ), f"Expected '/' separator in custom format, got {result[0]}"
-
-    def test_time_stamp_property_with_units(self):
-        """Verify time_stamp property returns dates when time has units.
-
-        Covers the delegation to get_time_variable().
-        """
-        nc = _make_nc_with_time_units(n_times=2)
-        result = nc.time_stamp
-        assert (
-            result is not None
-        ), "time_stamp should return dates when time units exist"
-        assert len(result) == 2, f"Expected 2 timestamps, got {len(result)}"
-
-
-class TestReadVariableFallbackPaths:
-    """Tests for _read_variable dimension indexing and classic mode."""
-
-    def test_read_variable_via_dimension_indexing(self):
-        """Verify _read_variable falls back to dimension indexing variable.
-
-        Covers when OpenMDArray returns None for a
-        dimension name, falls back to dim.GetIndexingVariable().
-        """
-        nc = make_2d_nc()
-        # Patch OpenMDArray to return None for the dimension variable
-        original_rg = nc._raster.GetRootGroup()
-
-        class PatchedRG:
-            """A wrapper that forces OpenMDArray to return None for 'x'."""
-
-            def __init__(self, real_rg):
-                """Store the real root group."""
-                self._real_rg = real_rg
-
-            def __getattr__(self, name):
-                """Delegate all calls except OpenMDArray."""
-                return getattr(self._real_rg, name)
-
-            def OpenMDArray(self, var_name, options=None):
-                """Return None for 'x' to force dimension indexing fallback."""
-                if var_name == "x":
-                    return None
-                if options is not None:
-                    return self._real_rg.OpenMDArray(var_name, options)
-                return self._real_rg.OpenMDArray(var_name)
-
-        with patch.object(
-            nc._raster,
-            "GetRootGroup",
-            return_value=PatchedRG(original_rg),
-        ):
-            result = nc._read_variable("x")
-        assert result is not None, "Should read 'x' via dimension indexing variable"
-        assert isinstance(
-            result, np.ndarray
-        ), f"Expected np.ndarray, got {type(result)}"
-
-    def test_read_variable_classic_mode_success(self):
-        """Verify _read_variable reads data in classic mode.
-
-        Covers the classic-mode path opening via
-        NETCDF:file:var string.
-        """
-        nc = NetCDF.read_file(
-            "tests/data/netcdf/noah-precipitation-1979.nc",
-            open_as_multi_dimensional=False,
-        )
-        # In classic mode, variables are Band1, Band2, etc.
-        # Try reading lon/lat which exist in the file
-        result = nc._read_variable("Band1")
-        assert result is not None, "Should read 'Band1' variable in classic mode"
-        assert isinstance(
-            result, np.ndarray
-        ), f"Expected np.ndarray, got {type(result)}"
-
-
-class TestGetVariableYFlipAndErrors:
-    """Tests for get_variable Y-flip correction and error paths."""
-
-    def test_get_variable_with_y_flip(self):
-        """Verify get_variable handles south-to-north Y orientation.
-
-        Covers the gt[5] > 0 correction branch
-        in get_variable.
-        """
-        # Create an MDIM dataset where lat is stored south-to-north
-        src = gdal.GetDriverByName("MEM").CreateMultiDimensional("yflip_test")
-        rg = src.GetRootGroup()
-        dtype = gdal.ExtendedDataType.Create(gdal.GDT_Float64)
-
-        # Create x dimension
-        dim_x = rg.CreateDimension("x", "HORIZONTAL_X", None, 5)
-        x_vals = rg.CreateMDArray("x", [dim_x], dtype)
-        x_vals.Write(np.arange(5, dtype=np.float64) + 0.5)
-        dim_x.SetIndexingVariable(x_vals)
-
-        # Create y dimension stored south-to-north (ascending)
-        dim_y = rg.CreateDimension("y", "HORIZONTAL_Y", None, 4)
-        y_vals = rg.CreateMDArray("y", [dim_y], dtype)
-        y_vals.Write(np.array([0.5, 1.5, 2.5, 3.5], dtype=np.float64))
-        dim_y.SetIndexingVariable(y_vals)
-
-        # Create data variable
-        data_arr = rg.CreateMDArray("temp", [dim_y, dim_x], dtype)
-        data_arr.Write(np.random.RandomState(88).rand(4, 5).astype(np.float64))
-        data_arr.SetNoDataValueDouble(-9999.0)
-
-        nc = Container(src)
-        var = nc.get_variable("temp")
-        # The Y-flip correction should have been applied
-        gt = var._geotransform
-        assert gt[5] <= 0, f"After Y-flip, gt[5] should be <= 0, got {gt[5]}"
-
-    def test_get_variable_classic_open_returns_none(self):
-        """Verify get_variable raises ValueError when gdal.Open returns None.
-
-        Covers the branch where gdal.Open returns None
-        in classic mode. GDAL sometimes returns None instead of raising.
-        """
-        nc = NetCDF.read_file(
-            "tests/data/netcdf/noah-precipitation-1979.nc",
-            open_as_multi_dimensional=False,
-        )
-        original_names = nc.variable_names[:]
-        nc._cached_variables = None
-        with (
-            patch.object(
-                nc,
-                "_get_variable_names",
-                return_value=original_names + ["fake_var"],
-            ),
-            patch("pyramids.netcdf.netcdf.gdal.Open", return_value=None),
-        ):
-            with pytest.raises(ValueError, match="Could not open variable"):
-                nc.get_variable("fake_var")
-
-    def test_get_variable_band_dim_read_error(self):
-        """Verify get_variable falls back to range when ReadAsArray fails.
-
-        Covers the RuntimeError branch where
-        ReadAsArray on the band dim indexing variable fails.
-        """
-        nc = _make_3d_nc()
-
-        original_read_md = nc._read_md_array
-
-        def patched_read_md(variable_name, x_dim=None, y_dim=None):
-            """Patch _read_md_array to return objects that simulate failure."""
-            src, md_arr, rg, ix, iy = original_read_md(
-                variable_name, x_dim=x_dim, y_dim=y_dim
-            )
-
-            # Wrap md_arr so GetDimensions returns a band dim whose
-            # indexing variable's ReadAsArray raises RuntimeError
-            class PatchedDim:
-                """Dimension wrapper that simulates ReadAsArray failure."""
-
-                def __init__(self, real_dim, should_fail):
-                    """Store the real dimension."""
-                    self._real = real_dim
-                    self._should_fail = should_fail
-
-                def GetName(self):
-                    """Return the dimension name."""
-                    return self._real.GetName()
-
-                def GetSize(self):
-                    """Return the dimension size."""
-                    return self._real.GetSize()
-
-                def GetIndexingVariable(self):
-                    """Return a variable whose ReadAsArray fails."""
-                    if self._should_fail:
-                        mock_iv = MagicMock()
-                        mock_iv.ReadAsArray.side_effect = RuntimeError(
-                            "simulated string variable"
-                        )
-                        return mock_iv
-                    return self._real.GetIndexingVariable()
-
-            class PatchedMDArr:
-                """MDArray wrapper with patched GetDimensions."""
-
-                def __init__(self, real_md_arr):
-                    """Store the real MDArray."""
-                    self._real = real_md_arr
-
-                def __getattr__(self, name):
-                    """Delegate all except GetDimensions."""
-                    return getattr(self._real, name)
-
-                def GetDimensions(self):
-                    """Return dims with patched band dim."""
-                    real_dims = self._real.GetDimensions()
-                    result_dims = []
-                    spatial = {len(real_dims) - 1, len(real_dims) - 2}
-                    for i, d in enumerate(real_dims):
-                        if i not in spatial:
-                            result_dims.append(PatchedDim(d, True))
-                        else:
-                            result_dims.append(PatchedDim(d, False))
-                    return result_dims
-
-            return src, PatchedMDArr(md_arr), rg, ix, iy
-
-        with patch.object(nc, "_read_md_array", side_effect=patched_read_md):
-            var = nc.get_variable("temperature")
-        # Should have fallen back to range-based values
-        assert (
-            var._band_dim_values is not None
-        ), "band_dim_values should be set (fallback to range)"
-        assert var._band_dim_values == [
-            0,
-            1,
-            2,
-        ], f"Expected [0, 1, 2] as range fallback, got {var._band_dim_values}"
-
-    def test_get_variable_md_arr_none(self):
-        """Verify get_variable handles case when md_arr is None.
-
-        Covers the branch where md_arr is None after
-        _read_md_array returns a non-MDArray (e.g. string type).
-        """
-        nc = _make_3d_nc()
-        # If _read_md_array returns None md_arr (second element), the code
-        # sets md_arr = None, then the fallback path sets defaults
-        original_read = nc._read_md_array
-
-        def patched_read(variable_name, x_dim=None, y_dim=None):
-            """Return None for md_arr to trigger default dim info."""
-            src, _, rg_ref, ix, iy = original_read(
-                variable_name, x_dim=x_dim, y_dim=y_dim
-            )
-            return src, None, rg_ref, ix, iy
-
-        with patch.object(nc, "_read_md_array", side_effect=patched_read):
-            var = nc.get_variable("temperature")
-
-        assert (
-            var._md_array_dims == []
-        ), f"Expected empty md_array_dims, got {var._md_array_dims}"
-        assert (
-            var._band_dim_name is None
-        ), f"Expected None band_dim_name, got {var._band_dim_name}"
-        assert (
-            var._band_dim_values is None
-        ), f"Expected None band_dim_values, got {var._band_dim_values}"
-        assert (
-            var._variable_attrs == {}
-        ), f"Expected empty variable_attrs, got {var._variable_attrs}"
-
-
-class TestGetVariableNonDataset:
-    """Tests for get_variable when _read_md_array returns non-Dataset."""
-
-    def test_get_variable_1d_string_returns_md_arr(self):
-        """Verify get_variable handles non-Dataset result from _read_md_array.
-
-        Covers the else branch where src from
-        _read_md_array is not a gdal.Dataset (e.g. string-type MDArray),
-        and cube is set to src directly.
-        """
-        # Create a dataset with a 1D string variable as a "data variable"
-        src = gdal.GetDriverByName("MEM").CreateMultiDimensional("str_var_test")
-        rg = src.GetRootGroup()
-        dtype = gdal.ExtendedDataType.Create(gdal.GDT_Float64)
-
-        # Create x and y dimensions (needed so they get excluded)
-        dim_x = rg.CreateDimension("x", "HORIZONTAL_X", None, 5)
-        x_vals = rg.CreateMDArray("x", [dim_x], dtype)
-        x_vals.Write(np.arange(5, dtype=np.float64))
-        dim_x.SetIndexingVariable(x_vals)
-
-        dim_y = rg.CreateDimension("y", "HORIZONTAL_Y", None, 4)
-        y_vals = rg.CreateMDArray("y", [dim_y], dtype)
-        y_vals.Write(np.arange(4, dtype=np.float64))
-        dim_y.SetIndexingVariable(y_vals)
-
-        # Create a 1D string array as data variable
-        str_dim = rg.CreateDimension("labels_dim", None, None, 3)
-        str_dtype = gdal.ExtendedDataType.CreateString()
-        str_arr = rg.CreateMDArray("labels", [str_dim], str_dtype)
-
-        nc = Container(src)
-        assert (
-            "labels" in nc.variable_names
-        ), f"'labels' should be a variable, got {nc.variable_names}"
-        var = nc.get_variable("labels")
-        # The result should be the MDArray itself (not a Dataset)
-        assert var is not None, "Variable should not be None"
-        assert var._is_subset is True, "Should be marked as subset"
-
-
-class TestGetVariableMultipleBandDims:
-    """Tests for get_variable with multiple non-spatial dims (issue #311).
-
-    Pre-#311 the build path discarded both band dims when there was more
-    than one non-spatial dim, leaving ``_band_dim_name = None`` and
-    locking out ``sel()``. Post-#311 the build path tracks every
-    non-spatial dim in ``_band_dim_names`` / ``_band_dim_values_map`` /
-    ``_band_dim_sizes``, while the legacy ``_band_dim_name`` /
-    ``_band_dim_values`` keep pointing at the first non-spatial dim.
-    """
-
-    def test_get_variable_with_two_band_dims(self):
-        """Build path tracks both band dims for a 4-D variable.
-
-        Test scenario:
-            Construct an in-memory 4-D MDArray with dims (time, ensemble,
-            y, x). Both non-spatial dims must appear in ``_band_dim_names``
-            in storage order; their sizes and coord values must be
-            populated; legacy fields point at the primary (time) dim.
-        """
-        src = gdal.GetDriverByName("MEM").CreateMultiDimensional("multi_band_dims")
-        rg = src.GetRootGroup()
-        dtype = gdal.ExtendedDataType.Create(gdal.GDT_Float64)
-
-        dim_x = rg.CreateDimension("x", "HORIZONTAL_X", None, 3)
-        x_v = rg.CreateMDArray("x", [dim_x], dtype)
-        x_v.Write(np.array([0.5, 1.5, 2.5]))
-        dim_x.SetIndexingVariable(x_v)
-
-        dim_y = rg.CreateDimension("y", "HORIZONTAL_Y", None, 3)
-        y_v = rg.CreateMDArray("y", [dim_y], dtype)
-        y_v.Write(np.array([2.5, 1.5, 0.5]))
-        dim_y.SetIndexingVariable(y_v)
-
-        dim_t = rg.CreateDimension("time", "TEMPORAL", None, 2)
-        t_v = rg.CreateMDArray("time", [dim_t], dtype)
-        t_v.Write(np.array([0.0, 1.0]))
-        dim_t.SetIndexingVariable(t_v)
-
-        dim_e = rg.CreateDimension("ensemble", None, None, 2)
-        e_v = rg.CreateMDArray("ensemble", [dim_e], dtype)
-        e_v.Write(np.array([1.0, 2.0]))
-        dim_e.SetIndexingVariable(e_v)
-
-        data = rg.CreateMDArray("temp", [dim_t, dim_e, dim_y, dim_x], dtype)
-        data.Write(np.random.rand(2, 2, 3, 3).astype(np.float64))
-        data.SetNoDataValueDouble(-9999.0)
-
-        nc = Container(src)
-        var = nc.get_variable("temp")
-        assert var._band_dim_names == (
-            "time",
-            "ensemble",
-        ), f"Expected ('time', 'ensemble'), got {var._band_dim_names!r}"
-        assert var._band_dim_sizes == (
-            2,
-            2,
-        ), f"Expected sizes (2, 2), got {var._band_dim_sizes!r}"
-        assert var._band_dim_values_map["time"] == [
-            0.0,
-            1.0,
-        ], f"time values mismatch: {var._band_dim_values_map.get('time')!r}"
-        assert var._band_dim_values_map["ensemble"] == [1.0, 2.0], (
-            f"ensemble values mismatch: "
-            f"{var._band_dim_values_map.get('ensemble')!r}"
-        )
-        assert (
-            var._band_dim_name == "time"
-        ), f"legacy primary must be 'time', got {var._band_dim_name!r}"
-        assert var._band_dim_values == [
-            0.0,
-            1.0,
-        ], f"legacy primary values mismatch: {var._band_dim_values!r}"
-
-
-class TestGetVariableAttrException:
-    """Tests for get_variable GetAttributes exception."""
-
-    def test_get_variable_attr_read_error(self):
-        """Verify get_variable handles GetAttributes failure gracefully.
-
-        Covers the except Exception: pass block when
-        GetAttributes raises on the MDArray.
-        """
-        nc = _make_3d_nc()
-        original_read_md = nc._read_md_array
-
-        def patched_read(variable_name, x_dim=None, y_dim=None):
-            """Wrap md_arr with one that fails on GetAttributes."""
-            src, md_arr, rg_ref, ix, iy = original_read_md(
-                variable_name, x_dim=x_dim, y_dim=y_dim
-            )
-
-            class AttrFailMDArr:
-                """MDArray wrapper that fails on GetAttributes."""
-
-                def __init__(self, real_arr):
-                    """Store the real MDArray."""
-                    self._real = real_arr
-
-                def __getattr__(self, name):
-                    """Delegate everything except GetAttributes."""
-                    return getattr(self._real, name)
-
-                def GetAttributes(self):
-                    """Raise to simulate failure."""
-                    raise RuntimeError("Cannot read attributes")
-
-            return src, AttrFailMDArr(md_arr), rg_ref, ix, iy
-
-        with patch.object(nc, "_read_md_array", side_effect=patched_read):
-            var = nc.get_variable("temperature")
-        assert (
-            var._variable_attrs == {}
-        ), f"Expected empty attrs after exception, got {var._variable_attrs}"
-
-
-class TestSetVariableAttrWriteException:
-    """Tests for set_variable attribute Write exception."""
-
-    def test_set_variable_attr_write_failure_silenced(self):
-        """Verify set_variable silences exceptions in attribute Write.
-
-        Covers the except Exception: pass block
-        when CreateAttribute or Write raises.
-        """
-        nc = make_2d_nc()
-        ds = _make_dataset_2d()
-
-        # We need the CreateAttribute call to succeed but Write to fail
-        # We'll patch CreateAttribute to return a mock whose Write raises
-        original_set_variable = NetCDF.set_variable
-
-        def intercept_set_variable(self_nc, var_name, dataset, **kwargs):
-            """Call set_variable but with an attr that will fail on Write."""
-            # Use an attr dict with a special sentinel
-            kwargs["attrs"] = {"will_fail": object()}
-            original_set_variable(self_nc, var_name, dataset, **kwargs)
-
-        # Simply test that the exception is silenced
-        # Using object() as attr value forces str() conversion in the
-        # else branch. The Write may or may not fail, but the test
-        # verifies no exception escapes.
-        nc.set_variable(
-            "fail_attr_var",
-            ds,
-            attrs={"key": object()},
-        )
-        rg = nc._raster.GetRootGroup()
-        md_arr = rg.OpenMDArray("fail_attr_var")
-        assert md_arr is not None, "Variable should exist despite attr issues"
-
-
-class TestSetVariableNoDataException:
-    """Tests for set_variable no-data exception handling."""
-
-    def test_set_variable_no_data_float_conversion_error(self):
-        """Verify set_variable handles exception in SetNoDataValueDouble.
-
-        Covers the except pass block when
-        SetNoDataValueDouble raises.
-        """
-        nc = make_2d_nc()
-        ds = _make_dataset_2d()
-        # Set a no_data_value that can't be converted to float
-        ds._no_data_value = ["not_a_number"]
-        # This should not raise - the exception is silenced
-        nc.set_variable("tricky_var", ds)
-        rg = nc._raster.GetRootGroup()
-        md_arr = rg.OpenMDArray("tricky_var")
-        assert md_arr is not None, "Variable should still be created"
-
-
-class TestSetVariableAttrException:
-    """Tests for set_variable attribute exception silencing."""
-
-    def test_set_variable_with_attr_create_failure(self, monkeypatch):
-        """set_variable silences a CreateAttribute failure (the except-pass branch).
-
-        Force every ``MDArray.CreateAttribute`` call to raise, then write a
-        variable with attributes. The per-attribute helper (`_write_attrs`)
-        swallows the failure, so set_variable must still create the variable and
-        return without propagating the error.
-        """
-        nc = make_2d_nc()
-        ds = _make_dataset_2d()
-
-        def boom(*args, **kwargs):
-            raise RuntimeError("forced CreateAttribute failure")
-
-        monkeypatch.setattr(gdal.MDArray, "CreateAttribute", boom)
-
-        # Must NOT raise despite every attribute write failing.
-        nc.set_variable("attr_err_var", ds, attrs={"units": "K", "flag": 1})
-
-        rg = nc._raster.GetRootGroup()
-        md_arr = rg.OpenMDArray("attr_err_var")
-        assert md_arr is not None, "Variable should still be created despite attr failure"
-        attr_names = [a.GetName() for a in (md_arr.GetAttributes() or [])]
-        assert "units" not in attr_names, "the failed attribute must not be written"
-
-
-class TestReadMdArray1DNumeric:
-    """Tests for _read_md_array with 1D numeric variables."""
-
-    def test_read_md_array_1d_numeric_via_custom_ds(self):
-        """Verify _read_md_array returns AsClassicDataset for 1D numeric arrays.
-
-        Covers the 1D numeric branch. Creates a custom
-        MDIM dataset with a 1D numeric variable to test this path.
-        Note: AsClassicDataset(0, 1, rg) may fail on some GDAL versions
-        for truly 1D arrays. We test the code path is reached.
-        """
-        src = gdal.GetDriverByName("MEM").CreateMultiDimensional("test_1d")
-        rg = src.GetRootGroup()
-        dtype = gdal.ExtendedDataType.Create(gdal.GDT_Float64)
-        dim = rg.CreateDimension("z", None, None, 5)
-        z_vals = rg.CreateMDArray("z", [dim], dtype)
-        z_vals.Write(np.array([1.0, 2.0, 3.0, 4.0, 5.0]))
-        dim.SetIndexingVariable(z_vals)
-
-        # Create a 1D data variable (not a dimension coordinate)
-        profile = rg.CreateMDArray("profile", [dim], dtype)
-        profile.Write(np.array([10.0, 20.0, 30.0, 40.0, 50.0]))
-
-        nc = Container(src)
-        result = nc._read_md_array("profile")
-        assert result is not None, "Should return result for 1D numeric array"
-
-
-class TestCubeDimensionNames:
-    """`get_variable(...).dimension_names` must mirror the container's view.
-
-    Pre-fix the property returned `None` on a variable subset because the
-    classic-mode in-memory `Dataset` underlying the cube has no GDAL root
-    group. The cube's dim names live on `_md_array_dims` and are the right
-    fall-through. See `pyramids-h1-followup.md` in the earthly planning
-    docs for context.
-
-    Note: when pyramids has to y-flip the source on read (ascending-lat
-    inputs), the spatial dim is renamed to `subset_lat_*` on the cube.
-    Real-world CDS-Beta files (already north-down) keep the original
-    name. Tests use the era5 fixture for byte-equality and the synthetic
-    4-D fixture for shape / first-two-dims invariants.
-    """
-
-    def test_cube_lists_all_dims_in_storage_order_on_real_4d(self):
-        """4-D cube reports every dim in storage order (no y-flip rename here).
-
-        Test scenario:
-            era5 CDS-Beta pressure-levels file is already north-down so
-            pyramids takes the no-flip read path. The cube's
-            `dimension_names` should match the container's exactly.
-        """
-        nc = NetCDF.read_file(
-            "tests/data/netcdf/era5_cds_beta_t_pressure_levels_jan2022.nc"
-        )
-        var = nc.get_variable("t")
-        assert var.dimension_names == [
-            "valid_time",
-            "pressure_level",
-            "latitude",
-            "longitude",
-        ], f"got {var.dimension_names!r}"
-
-    def test_cube_dimension_names_matches_container_for_real_4d(self):
-        """Cube's `dimension_names` mirrors container's on a no-flip file."""
-        nc = NetCDF.read_file(
-            "tests/data/netcdf/era5_cds_beta_t_pressure_levels_jan2022.nc"
-        )
-        var = nc.get_variable("t")
-        assert (
-            var.dimension_names == nc.dimension_names
-        ), f"cube={var.dimension_names!r} container={nc.dimension_names!r}"
-
-    def test_cube_dimension_names_first_two_match_band_dims_on_synthetic(self):
-        """Synthetic 4-D cube reports both band dims first; spatial dims may
-        be renamed by pyramids' y-flip but count is still 4.
-        """
-        nc = NetCDF.read_file("tests/data/netcdf/pyramids-netcdf-4d.nc")
-        var = nc.get_variable("temperature")
-        names = var.dimension_names
-        assert names is not None, "cube dim names must not be None after fix"
-        assert len(names) == 4, f"4-D cube must have 4 dims, got {names!r}"
-        assert names[:2] == [
-            "time",
-            "pressure_level",
-        ], f"band dims must be first two, got {names!r}"
-
-    def test_cube_dimension_names_is_independent_copy(self):
-        """Mutating the returned list must not alter `_md_array_dims`."""
-        nc = NetCDF.read_file("tests/data/netcdf/pyramids-netcdf-4d.nc")
-        var = nc.get_variable("temperature")
-        names = var.dimension_names
-        original = list(var._md_array_dims)
-        names.append("bogus")
-        assert "bogus" not in var._md_array_dims, (
-            f"_md_array_dims was mutated through dimension_names: "
-            f"{var._md_array_dims!r}"
-        )
-        assert var._md_array_dims == original, (
-            "subsequent reads of dimension_names should still match the "
-            "original cached list"
-        )
-
-    def test_cube_with_no_md_array_dims_returns_none(self):
-        """Defensive: a cube whose `_md_array_dims` is empty returns `None`."""
-        nc = NetCDF.read_file("tests/data/netcdf/pyramids-netcdf-4d.nc")
-        var = nc.get_variable("temperature")
-        var._md_array_dims = []
-        assert (
-            var.dimension_names is None
-        ), f"empty cache should yield None, got {var.dimension_names!r}"
-
-    def test_container_dimension_names_unchanged(self):
-        """The container path is unchanged: still reads from the root group.
-
-        Test scenario:
-            On the bundled 4-D synthetic, the container reports the
-            exact dim names from the file (no y-flip rename, since the
-            container is the original MDIM dataset).
-        """
-        nc = NetCDF.read_file("tests/data/netcdf/pyramids-netcdf-4d.nc")
-        assert nc.dimension_names == [
-            "time",
-            "pressure_level",
-            "lat",
-            "lon",
-        ], f"container path regressed: {nc.dimension_names!r}"
